chore(markdown): hoist RenderHTML engine to app root
react-native-render-html's combined <RenderHTML> component mounts its
own TRenderEngineProvider + RenderHTMLConfigProvider per instance.
During streaming the lib's useProfiler fired "costly tree rerenders"
warnings on every token (engine being rebuilt at <50ms intervals);
the existing 150ms streaming throttle existed to stay above that
threshold but produced perceptibly bursty UX in exchange.

Per the library maintainer's canonical guidance (SO/68966121), split
the providers from the source: host the heavy engine + config once at
the app root and use the lightweight <RenderHTMLSource> in components.

- New `MarkdownProvider` wraps the app inside PaperProvider, owning:
  - TRenderEngineProvider: tagsStyles (theme-derived), systemFonts,
    customHTMLElementModels
  - RenderHTMLConfigProvider: defaultTextProps, renderers (including
    the CodeRenderer hoisted to its own file for module-stable identity)
- MarkdownView now renders only <RenderHTMLSource source contentWidth>.
  Per-token updates change just `source`; the engine stays mounted, so
  the lib's profiler no longer fires.
- ReasoningBlock is unchanged — it carries its own tagsStyles override
  and isn't in the streaming hot path; can be migrated later if needed.

Tests: the centralised render wrapper in jest/test-utils now mounts
MarkdownProvider inside PaperProvider so any RenderHTML-using
component test gets the ambient providers. MarkdownView's own test
switches from bare @testing-library/react-native to the custom render.

Net effect: streaming UI updates become cheap. The 150ms throttle is
no longer load-bearing — that decision will be revisited separately.
