+++ jest/test-utils.tsx
@@ -12,6 +12,7 @@ import {themeFixtures} from './fixtures/theme';
 
 import {user as userFixture} from './fixtures';
 
+import {MarkdownProvider} from '../src/components/MarkdownView';
 import {UserContext} from '../src/utils';
 import type {Theme} from '../src/utils/types';
 
@@ -49,8 +50,13 @@ const customRender = (
       withBottomSheetProviderWrapper
     );
 
+    // Sits inside PaperProvider so the MarkdownProvider can read theme,
+    // and provides the ambient TRenderEngineProvider/RenderHTMLConfigProvider
+    // any component using RenderHTMLSource (e.g. MarkdownView) depends on.
     const withPaperProvider = (
-      <PaperProvider theme={theme}>{withNavigationWrapper}</PaperProvider>
+      <PaperProvider theme={theme}>
+        <MarkdownProvider>{withNavigationWrapper}</MarkdownProvider>
+      </PaperProvider>
     );
 
     const withSafeAreaWrapper = withSafeArea ? (
