+++ src/components/MarkdownView/index.ts
@@ -1,1 +1,2 @@
 export * from './MarkdownView';
+export * from './MarkdownProvider';
